Reorganize HexView sources under workarea/hexview
