+++ src/ui/qt/workarea/hexview/HexView.cpp
