+++ src/ui/qt/workarea/hexview/HexView.h
