+++ src/ui/qt/workarea/hexview/HexViewInput.h
